Remove 'A' argument from dns.resolve4
According to the documentation, resolve4 already queries A records, and type is not one of the options.
I have been observing DNS timeouts after a few hundred milliseconds despite the configuration being 25s.
GPT suggested this is the cause and indeed forking the plugin with this fix causes the timeouts to stop.